new_version/backend/pyproject.toml
[project]
name = "bitrix-sync-backend"
version = "1.0.0"
description = "Bitrix24 CRM to PostgreSQL sync service"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "fast-bitrix24>=1.8.0",
    "pydantic-settings>=2.0.0",
    "supabase>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "apscheduler>=3.10.0",
    "tenacity>=8.2.0",
    "structlog>=24.0.0",
    "alembic>=1.13.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
